fetch/metadata/redirect/multiple-redirect-https-downgrade-upgrade.sub.html

<!DOCTYPE html>
<html>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=/fetch/metadata/resources/helper.js></script>
<script src=/fetch/metadata/resources/redirectTestHelper.sub.js></script>
<script src=/common/security-features/resources/common.sub.js></script>
<script src=/common/utils.js></script>
<body>
<script>
  let nonce = token();
  let expected = {"site": "cross-site", "user": "", "mode": "cors", "dest": "font"};

  // Validate various scenarios handle a request that redirects from https => http => https
  // correctly and avoids disclosure of any Sec- headers.
  RunCommonRedirectTests("Https downgrade-upgrade", MultipleRedirectTo, expected);
</script>
<script
    src="https://{{host}}:{{ports[https][0]}}/fetch/api/resources/redirect.py?location=http%3A%2F%2F{{host}}%3A{{ports[http][0]}}%2Ffetch%2Fapi%2Fresources%2Fredirect.py%3Flocation%3Dhttps%253A%252F%252F{{host}}%253A{{ports[https][0]}}%252Ffetch%252Fmetadata%252Fresources%252Fecho-as-script.py"></script>
<script>
  test(t => {
    t.add_cleanup(_ => {
      header = null;
    });
    assert_header_equals(header, {"site": "cross-site", "user": "", "mode": "no-cors", "dest": "script"},
        "Https downgrade-upgrade script => No headers");
  }, "Https downgrade-upgrade script => No headers");
</script>
</body>
